Extracted `mdbook init` to separate file/module
